Use the SYNOPSIS as the usage text
The default usage message is fine, unless the command is
expected to take arguments, in which case there is no
easy way to pass that information.  This patch looks in
the command class file to see if there's a SYNOPSIS,
and if there is one, use that as the usage text.

Note: since this is Xmas Eve, I only hacked a dirty proof
of concept. If you like the idea, that patch should be cleaned up
before it's fully merged in.