Add the ability to declare dependent options to MooseX::App
In many real-world scenarios, a option could require one or more other options to make your application work properly, for example if your application writes a configuration there may be an option to to specify the format of that configuration file. A new plugin, Requires, implements these dependent options. The implementation consists of two metaroles, an attribute that defines the requires key and a class that defines the logic for how options that specify the requires key behave.