[FIX] test uses recommened module
Text::WagnerFischer is only recommended, not required. The unit test
shows how to use it (as an example) but tries not to run the code that
uses Text::WagnerFischer when executed as part of the unit tests. For
this the test accidently checks $ENV{HARNESS} when the correct variable
is called $ENV{HARNESS_ACTIVE}. Fixed.