Update Base.pm
According to the documentation for MooseX::App, the ordering of options / parameters in the usage message is something that can be altered by specifying the 'cmd_position' attribute property, however it does not appear to be functional for options.  

I am not sure if this was intended and the documentation is incorrect, but this patch enables the feature for options.